zy cj Changed db field from integer to numeric

+++ db/scripts/posts/createTable.js
@@ -1,7 +1,7 @@
 import db from "../../connection.js";
 
 const response = await db.query(
-  `CREATE TABLE IF NOT EXISTS posts (id SERIAL PRIMARY KEY, title TEXT, full_name TEXT, description TEXT, location TEXT, phone_number TEXT, email TEXT, free BOOLEAN NOT NULL, price INTEGER, date TEXT);`
+  `CREATE TABLE IF NOT EXISTS posts (id SERIAL PRIMARY KEY, title TEXT, full_name TEXT, description TEXT, location TEXT, phone_number TEXT, email TEXT, free BOOLEAN NOT NULL, price NUMERIC, date TEXT);`
 );
 
 console.log(response);
